Extended seeds/way points and support for seeds on the edge of images.
Extended seeds can be useful in minimal path applications. For example
it might be known that a feature crosses a plane, but not exactly where.
This patch allows the user to find the minimal path between a point and
a plane, or between two planes.

There are issues with the back propogation steps. The original implementation
was returning a value of 0 from the image cost function when the queried location
was outside the image buffer. This causes problems whenever the path touches
an image edge (or when a seed is on an image edge), as the back propogation
process will terminate. This has been addressed by supplying methods to
set the values returned when the query is outside the image buffer.

Similar issues can occur when a speed image includes paths that are a single
voxel wide. Gradient descent back propogation can be incorrectly terminated
because the interpolator produces an artificial minima due to very large
values in the arrival time image around the single voxel path. This
problem hasn't been addressed.